Modified all Skin classes due to the modifications in JavaFX 8 b74 preview
registerChangeListener method was moved from SkinBase to BehaviorSkinBase. So I changed all Skins in the way that they now extend BehaviorSkinBase instead of SkinBase to be able to use the registerChangeListener methods in the Skin class. In the future the property change listeners seems to be forwarded to the Behavior to handle them there.